Replace schemas mixins
